Fix comparison report when one column is all NAs (#343)
* Fix compariosn report when one column is all NAs

With situations when one of the column in the comparison was
all NA's then this would break the reporting. For some reason
when the matching (boolean) of the actual and expected columns happened when there was a categorical value compared to a NA, the result was NA, rather than False, as it would happen for the other elements in the cols compared.
    
This has now been addressed at the intersect rows level which doesn't seem to break the reporting anymore.

* Fix column_equal to work with StringArrays with pd.NA values not returning booleans

* Add test for fn column_equal to work with StringArrays with pd.NA 

Add test for fn column_equal to work with StringArrays containing pd.NA values not returning booleans when compared with other df's with rows of StringArrays

* Fix linter error

Printing out the report would've been useful for this test, but looks like it makes the linter fail the build. This has now been fixed.

* Fix column_equal to work with StringArrays with pd.NA values

Fix column_equal to work with StringArrays with pd.NA values not returning booleans, and update formatting to match the linter expectation

* Add test for fn column_equal to work with StringArrays with pd.NA

Add test for fn column_equal to work with StringArrays containing pd.NA values not returning booleans when compared with other df's with rows of StringArrays, and format test to match the linter.